test_cases

+++ include/ConvexHull.h
@@ -2,7 +2,7 @@
 #define CONVEXHULL_H
 
 #include <bits/stdc++.h>
-#define Point pair<int, int>
+#define Point pair<long long int, long long int>
 using namespace std;
 
 class ConvexHull{
